+++ app/javascript/vue/tasks/nomenclature/filter/components/filters/status.vue
@@ -127,7 +127,9 @@ export default {
 
     nomenclatureCode: {
       handler () {
-        this.merge()
+        if (Object.keys(this.statusList).length) {
+          this.merge()
+        }
       }
     }
   },